// src/pages/privacy_policy/types.ts
export interface PrivacyPolicy {
  title: string;
  sub_title: string;
  description: string;
  list_title: string;
  list: string[];
  sub_list_title: string;
  sub_list: string[];
  how_to_use_title: string;
  how_to_use_description: string;
  links_title: string;
  links_description: string;
  security_title: string;
  security_description: string;
  personal_info_title: string;
  personal_info_sub_title: string;
  personal_info_list: string[];
  personal_info_description: string;
  personal_info_sub_description: string;
  concat_title: string;
  contact_description: string;
  email: string;
  address: string;
  phone: string;
  hint_question: string;
}